feature/86ep326fg #Code refactoring hightlight setState function and wordsRendering inside hcp utils
