+++ apps/simulator/src/core/utils/hcpUtils.tsx
@@ -1,3 +1,5 @@
+import { Dispatch, SetStateAction } from 'react';
+
 export type selectedWordType = {
     word: string
     wordIndex: number
@@ -78,5 +80,51 @@ export const getHighlightedValues = (wordsInItem: string[]) => {
 
         return filterBySelectedAriaLabel
     }
+
 }
 
+
+export const handleHighlight = (
+    wordsInItem: string[],
+    setHighlightedWords: Dispatch<SetStateAction<selectedWordType[]>>
+) => {
+    const handleMouse = () => {
+        const highlightedValue = getHighlightedValues(wordsInItem)
+
+        if (highlightedValue) {
+            setHighlightedWords((prevValues) => {
+                const combinedValues = [...prevValues, ...highlightedValue]
+                const removeDuplication = removeArrayDuplication(combinedValues)
+
+                const removeRehighlighted = prevValues.map((value) => {
+                    const filterByValues =
+                        highlightedValue.map((f) => {
+                            if (f.wordIndex === value.wordIndex) {
+                                return f.wordIndex
+                            }
+                        }).filter((filteredValue) => filteredValue !== undefined).pop()
+                    return filterByValues
+                }).filter((filteredValue) => filteredValue !== undefined)
+
+                const filterCombinedValues = combinedValues.filter((values) => !removeRehighlighted.includes(values.wordIndex))
+                return filterCombinedValues
+                return removeDuplication
+            })
+        }
+    }
+    document.addEventListener("mouseup", handleMouse);
+    return () => {
+        document.removeEventListener("mouseup", handleMouse);
+    };
+}
+
+export const renderHighlightText = (values: string[], highlightedWords: selectedWordType[]) => {
+    const highlightedWordIndices = highlightedWords.map((w) => w.wordIndex)
+    return values.map((word, index) => {
+        const isHighlighted = highlightedWordIndices.includes(index)
+        const convertNewLine = word.replace(/\n/g, '<br/> ')
+        const theObj = { __html: convertNewLine + ' ' }
+
+        return word.includes('<br/>') ? <span key={index} dangerouslySetInnerHTML={theObj} /> : <span key={index} className={`${isHighlighted ? 'bg-yellow-300' : ''}`} aria-label={`${index}`} dangerouslySetInnerHTML={theObj} />
+    })
+}
